autism/backend/AI_Server/requirements.txt
absl-py==2.3.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
arabic-reshaper==3.0.0
attrs==25.3.0
av==14.4.0
bcrypt==4.3.0
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.6.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==44.0.2
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.6.1
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.1
fastapi==0.104.1
fastapi-cli==0.0.4
Flask==3.1.0
flask-cors==6.0.0
flatbuffers==25.2.10
fonttools==4.53.0
graphviz==0.20.3
greenlet==3.2.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
Jinja2==3.1.4
jwt==1.3.1
kiwisolver==1.4.5
mangum==0.17.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
motor==3.7.1
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.2
opt_einsum==3.4.0
orjson==3.10.5
packaging==24.1
pandas==2.1.4
passlib==1.7.4
pillow==10.3.0
playwright==1.52.0
protobuf==4.25.8
pyasn1==0.6.1
pycparser==2.22
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyee==13.0.0
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.13.2
PyMuPDF==1.24.5
pymupdf-fonts==1.0.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
PyPDF2==3.0.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.15.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.27.0
tinycss2==1.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.24.0
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
Werkzeug==3.1.3
gunicorn==21.2.0
